Add files via upload
The sinking of the RMS Titanic is one of the most infamous shipwrecks in history. In this project, we utilize the Titanic dataset to build a predictive model that determines the survival of passengers based on their characteristics. This project is a classic example for beginners in machine learning and data science.